expand: change to BUILDDIR before running bitbake stuff
Without this change, we get this error:

The BBPATH variable is not set and bitbake did not find a
conf/bblayers.conf file in the expected location.
Maybe you accidentally invoked bitbake from the wrong directory?

With this change `ye expand` can be run from any directory, as long as
BUILDDIR is set.

Change-Id: Ie66c40f4db58a13d0c74f9339f29ce445bd9bb1f
Signed-off-by: Mario Domenech Goulart <mario@ossystems.com.br>